requirements.txt
python 3.13.3
pygame-ce 2.5.6 
numpy 2.3.1